Consolidated switching projects behavior into the ValueChanged callback.
Now everything on the Import tab related to switching projects is under the projectNameFieldValueChanged function.
The pgui itself still takes care of initial startup like looking for the most recently used project name on first startup. But that's just to get the project name, it then quickly hands off the management of project metadata to the callback function.