Improving data types handling
Now imports can be done in a similar way as processing functions. Meaning, every import will have a method number & letter. The method number corresponds to the actual code used to import. This may vary based on data file types, etc. but will NOT change where the data is stored or how it looks as a final product.
The letter corresponds to any metadata that the user wants to enter about that data type. For example, mocap & FP data might specify axes orientations, and EMG might specify the locations of electrode placements, etc.