Finished handling data types switching
Now recalls which data type was just used last when opening again. And the method number/letter properly associates with the correct data type, even if it shares characters with another data type.
Untested, but this would probably even work with the same data type imported multiple ways! Though I'd have to change the "add data type" button to allow that.
Also, added a fun "Let's Get Ready to Rumble!" sound bite that plays the very first time that the app is loaded.