First pass at Process > Run code. Currently only trial level written, all untested.
Realized that I need to specify all of the processing levels used in a particular function in order to appropriately create a new function from a template.
The choice of template should consider the highest & lowest processing levels. e.g. a 'P' & 'T' function will start at project level, iterate through subjects, and then also through trials.
In terms of where the function is called, only the lowest level matters. e.g. a 'P' & 'T' function will be run at the function level so that the trial-level inputs can update. The project level inputs will be repeatedly called, but that should be ok.