Decided on data input method & created lots of processing templates
Was previously struggling with how exactly to input all arguments from args function, but I figured it out! When creating new function, user specifies the levels of Input & Output variables and it is created from that template. Then when running the function, the higher level of those two specifications is used to call the function. If a lower level input argument is needed, it is called within the processing function (included in template).