Put old versions into folder. Tried to handle indexed input arguments in getArg, needs to be finished
Non-scalar input arguments can be structs, cells, or any other type of array. Need to handle that in getArg (left side of the equals sign).
getArg can currently handle eval'ing manual input on the right side of the equals sign, but it cannot currently handle eval'ing manual input in combination with a projectStruct input (e.g. with multiplication). To do in the future!