Updated Import_argsTemplate to use functions
Instituted checks to ensure that ONLY local functions defined in the current args file are possibly implemented during that args call.
Question 1: Do I need an explicit argument to indicate that this is input or output arg call?
Question 2: Do I need to pass the output variable to the args function for some minor manipulation? Or can I assume that 100% of that is done within the import/process functions themselves?