Fixed computer ID issue! Now propagating that change & Projects tab
If project is not set up, user can't navigate to other tabs (for simplicity of setup, similar to how DLC does it). As the project is set up, the appropriate Projects tab components become visible one at a time, until the project is "fully set up" and the other tabs become available.