Function run order is now set per-node per-split.
Arguments are no longer removed when disconnecting nodes! This is in case it's a mistake, they will re-populate when the user re-connects the nodes. However, this may cause problems when attempting to delete a variable or processing split that is still present in these "shadow splits" (i.e. is present in the variable but is not one of the visible edges). I'll cross that bridge when I get there, I should just have to do a check to see if it is one of the visible edges.