readArgs fcn can read the getArg and setArg arguments to propagate into the GUI before vars are assigned.
The idea is that when adding a pre-existing function to the GUI, the input & output arguments will be read in and propagate into the input & output vars sections in the fcnArgsUITree. Then, when adding vars to the fcn, the user will add the "name in GUI" directly to a "name in code", bypassing the need for any typing. If no getArg/setArg vars already exist in the fcn, then it's blank and vars are added as normal, with the name in code needing to be typed in. If vars in the fcn change, then changes propagate to the GUI too.