Added my projectIndependentSettings back in after creating release.
To create a release, because the settings are stored IN the app's folder, I have to move the settings folder elsewhere to do the release, then move it back after creating the release. Not the worst, but kind of builds on the idea that maybe I don't want to have a file people need to interact with in the app's folder?
Pros: It's probably not disturbing anything else, can be kept under source control.
Cons: The source control would likely affect people's ability to contribute to the project (like it is for me), I don't want people to have to go into the app's folders if they don't want to.