Removed "fig" requirement from main helper functions.
This allows the code to be split into two: functions that manipulate the figure objects take in the fig as an input, and things that strictly manage files (no PGUI graphics objects) do not need them as an input at all. This will help tremendously for a run code.

Here I removed the "fig" inputs from all of the functions that I could, though there's likely a straggler or two. In the future I should split every callback into the same two categories and make subfunctions for the non-PGUI aspects. This will make the non-GUI and GUI code the same!