Version comparison is TOUGH. Getting too many functions, and then how to sort them??
Decided to use the "InputOfProcess"/"OutputOfProcess" and "InputVaiables"/"OutputVariables" fields that are already present. Not as versatile between classes but that's ok for now because it already robustly exists.

Also I'm getting too many processing functions because if one input variable is used way later in the pipeline, the algorithm currently has no way to know that that function is not needed. For example, "TBCMPosition" function was detecting "Gap_Fill_Foot_Markers" as a dependency because they both rely on smoothed marker data. But clearly "Gap_Fill_Foot_Markers" is not a dependency for TBCMPosition.