Close to being able to compare versions! Just need to order the functions.
I got rid of the extraneous functions, but now I'm struggling to order the functions. Starting with the first function, then moving forward making sure that each function's input variables (that are not hard-coded) are all covered by the prior functions' output variables before being placed as the next link in the chain.

Stupid issue right now is that the input & output variables are not staying attached to the function names.