Improving object loading robustness (in progress)
loadJSON should check that all fieldnames are present (and valid? Much more difficult!) before proceeding. Uses "createNewObject" to do that.

This/similar pipeline will also be used when copying objects to a new version. For that, everything in the JSON stays the same except the UUID? What about variable versions?