Linkage struct reading/writing, checking struct fields.
Linkage struct needs to be "translated" when reading/writing it because I want MATLAB to use an Nx3 matrix, but JSON doesn't support that directly. Instead, to write the file I convert it in MATLAB to an Nx1 cell array with each cell being a 1x3. Then to read it I do the reverse operation.

Also checking fields, ensuring that all fields present in a newly generated object (as a template) are also present in the file being loaded. Returns the modified struct.