Added "out of date" checkbox, and improved the speed of the ordering algorithm!
The ordering algorithm currently requires all edges to be found going into each node. However, I need to improve the algorithm to handle when only one edge is found going into the node. This is because sometimes I only care about if any edge at all goes into a node (e.g. if one input variable to a function is out of date, then the function and all its output variables are out of date).